+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # 0.0.2
-- 
+- backend ticker data with yfinance
+- beautiful UI, ticker chart for selected company 
+- basic chat streaming with langgraph agent workflow
 
 # 0.0.1
 - git init, 3 main views in next.js, dpeloyment on vercel